change some repository_storage_service enpoints to appropriate REST method

+++ services/repository_storage_service/src/repository_storage_service/main.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from fastapi import FastAPI
 import uvicorn
 from git import Repo
+from pydantic import BaseModel
 
 from repository_storage_service.utils import get_fs_object_content
 
@@ -13,11 +14,15 @@ app = FastAPI()
 os.chdir('./services/repository_storage_service/') # change current working directory
 os.makedirs('./repositories', exist_ok=True)
 
+class CloneRepositoryRequest(BaseModel):
+    repository_url: str
+
 # This endpoint should be called only by the scheduler service, which will provide the repository URL to be cloned.
 # The scheduler service will be responsible for managing the list of repositories to be cloned and for calling this endpoint with the appropriate repository URL.
 # For now this is a get endpoint that takes the repository URL as a query parameter, but in the future it should be a post endpoint that takes the repository URL in the request body.
-@app.get('/clone')
-def clone_repository(repository_url: str):
+@app.post('/clone_repository')
+def clone_repository_endpoint(clone_repository_request: CloneRepositoryRequest):
+    repository_url: str = clone_repository_request.repository_url
     destination = './repositories/' + repository_url.split('/')[-1].replace('.git', '')
     if Path(destination).exists():
         # fetch the latest changes if the repository already exists
@@ -37,7 +42,7 @@ def clone_repository(repository_url: str):
 # This endpoint should also be called by the dependency analyzer service (with display_files_content set to True) to get the content of the important manifest files in the repository to then parse them and get the dependencies in the repository.
 # This endpoint might be called by the dependency modifier service (with display_files_content set to True) to get the content of the important manifest files in the repository to then modify them and write them back to the repository. (i might also remove the Dependency modifier service and let this Repository storage service handle the modification directly through the MCP call)
 @app.get('/repositories/{path:path}')
-def get_fs_object_content(path: str, display_files_content: bool = False):
+def get_fs_object_content_endpoint(path: str, display_files_content: bool = False):
     fs_object_path = Path(f'./repositories/{path}')
     if not fs_object_path.exists():
         return { 'error': 'File or directory not found' }
@@ -46,16 +51,22 @@ def get_fs_object_content(path: str, display_files_content: bool = False):
     return fs_object_content
 
 
+class UpdateFileContentRequest(BaseModel):
+    new_content: str
+
 # This endpoint should be called by the Dependency modifier service (which is initially called by the LLM through the MCP server)
 # note: the Dependency modifier service will be responsible for providing correct content to be written in the file
 @app.put('/repositories/{path:path}')
-def update_file_content(path: str, new_content: str):
+def update_file_content_endpoint(path: str, update_file_content_request: UpdateFileContentRequest):
     file_path = Path(f'./repositories/{path}')
     if not file_path.exists():
         return { 'error': 'File not found' }
 
+    if not file_path.is_file():
+        return { 'error': 'Path is not a file' }
+
     with open(file_path, 'w') as f:
-        f.write(new_content)
+        f.write(update_file_content_request.new_content)
 
     return { 'message': f'File {path} updated successfully.' }
